refactor moveplayer
